refactor psbt_creator and utxo_scanner into commands (pattern)
